split padding tests into a new file
