+++ tests/test_sha256_hashing.py
@@ -78,32 +78,3 @@ class Test_Hashing(unittest.TestCase):
         self.assertEqual(hash_known, hash_file)
         remove("test_file.tmp")
 
-
-class Test_Padding(unittest.TestCase):
-    def setUp(self):
-        from sha256 import pad_message
-
-        self.pad_message = pad_message
-
-    def test_padding(self):
-        message = bytearray("abcd", "utf-8")
-        padded_message = self.pad_message(message)
-        message += b"\x80" + b"\x00" * 58 + b"\x20"
-        self.assertEqual(padded_message, message)
-
-    def test_padding_for_empty(self):
-        message = bytearray()
-        padded_message = self.pad_message(message)
-        message += b"\x80" + b"\x00" * 63
-        self.assertEqual(padded_message, message)
-
-    def test_padding_for_big(self):
-        message = bytearray("a" * 534, "utf-8")
-        padded_message = self.pad_message(message)
-        message += b"\x80" + b"\x00" * 33 + b"\x00\x00\x00\x00\x00\x00\x10\xb0"
-        self.assertEqual(padded_message, message)
-
-    def test_padding_for_length(self):
-        message = bytearray("a" * 534, "utf-8")
-        padded_message = self.pad_message(message)
-        self.assertEqual(len(padded_message), 576)
